+++ helper/schema/schema.go
@@ -205,7 +205,7 @@ func (m schemaMap) Diff(
 	}
 
 	for k, schema := range m {
-		err := m.diff(k, schema, result, d)
+		err := m.diff(k, schema, result, d, false)
 		if err != nil {
 			return nil, err
 		}
@@ -225,7 +225,7 @@ func (m schemaMap) Diff(
 
 		// Perform the diff again
 		for k, schema := range m {
-			err := m.diff(k, schema, result2, d)
+			err := m.diff(k, schema, result2, d, false)
 			if err != nil {
 				return nil, err
 			}
@@ -427,7 +427,8 @@ func (m schemaMap) diff(
 	k string,
 	schema *Schema,
 	diff *terraform.InstanceDiff,
-	d *ResourceData) error {
+	d *ResourceData,
+	all bool) error {
 	var err error
 	switch schema.Type {
 	case TypeBool:
@@ -435,13 +436,13 @@ func (m schemaMap) diff(
 	case TypeInt:
 		fallthrough
 	case TypeString:
-		err = m.diffString(k, schema, diff, d)
+		err = m.diffString(k, schema, diff, d, all)
 	case TypeList:
-		err = m.diffList(k, schema, diff, d)
+		err = m.diffList(k, schema, diff, d, all)
 	case TypeMap:
-		err = m.diffMap(k, schema, diff, d)
+		err = m.diffMap(k, schema, diff, d, all)
 	case TypeSet:
-		err = m.diffSet(k, schema, diff, d)
+		err = m.diffSet(k, schema, diff, d, all)
 	default:
 		err = fmt.Errorf("%s: unknown type %#v", k, schema.Type)
 	}
@@ -453,7 +454,8 @@ func (m schemaMap) diffList(
 	k string,
 	schema *Schema,
 	diff *terraform.InstanceDiff,
-	d *ResourceData) error {
+	d *ResourceData,
+	all bool) error {
 	o, n, _, computedList := d.diffChange(k)
 	nSet := n != nil
 
@@ -481,7 +483,7 @@ func (m schemaMap) diffList(
 	// If the new value was set, and the two are equal, then we're done.
 	// We have to do this check here because sets might be NOT
 	// reflect.DeepEqual so we need to wait until we get the []interface{}
-	if nSet && reflect.DeepEqual(os, vs) {
+	if !all && nSet && reflect.DeepEqual(os, vs) {
 		return nil
 	}
 
@@ -502,7 +504,7 @@ func (m schemaMap) diffList(
 	// If the counts are not the same, then record that diff
 	changed := oldLen != newLen
 	computed := oldLen == 0 && newLen == 0 && schema.Computed
-	if changed || computed {
+	if changed || computed || all {
 		countSchema := &Schema{
 			Type:     TypeInt,
 			Computed: schema.Computed,
@@ -539,7 +541,7 @@ func (m schemaMap) diffList(
 		// just diff each.
 		for i := 0; i < maxLen; i++ {
 			subK := fmt.Sprintf("%s.%d", k, i)
-			err := m.diff(subK, &t2, diff, d)
+			err := m.diff(subK, &t2, diff, d, all)
 			if err != nil {
 				return err
 			}
@@ -549,7 +551,7 @@ func (m schemaMap) diffList(
 		for i := 0; i < maxLen; i++ {
 			for k2, schema := range t.Schema {
 				subK := fmt.Sprintf("%s.%d.%s", k, i, k2)
-				err := m.diff(subK, schema, diff, d)
+				err := m.diff(subK, schema, diff, d, all)
 				if err != nil {
 					return err
 				}
@@ -566,8 +568,8 @@ func (m schemaMap) diffMap(
 	k string,
 	schema *Schema,
 	diff *terraform.InstanceDiff,
-	d *ResourceData) error {
-	//elemSchema := &Schema{Type: TypeString}
+	d *ResourceData,
+	all bool) error {
 	prefix := k + "."
 
 	// First get all the values from the state
@@ -590,7 +592,7 @@ func (m schemaMap) diffMap(
 		old := stateMap[k]
 		delete(stateMap, k)
 
-		if old == v {
+		if old == v && !all {
 			continue
 		}
 
@@ -613,15 +615,35 @@ func (m schemaMap) diffSet(
 	k string,
 	schema *Schema,
 	diff *terraform.InstanceDiff,
-	d *ResourceData) error {
-	return m.diffList(k, schema, diff, d)
+	d *ResourceData,
+	all bool) error {
+	if !all {
+		// This is a bit strange, but we expect the entire set to be in the diff,
+		// so we first diff the set normally but with a new diff. Then, if
+		// there IS any change, we just set the change to the entire list.
+		tempD := new(terraform.InstanceDiff)
+		tempD.Attributes = make(map[string]*terraform.ResourceAttrDiff)
+		if err := m.diffList(k, schema, tempD, d, false); err != nil {
+			return err
+		}
+
+		// If we had no changes, then we're done
+		if tempD.Empty() {
+			return nil
+		}
+	}
+
+	// We have changes, so re-run the diff, but set a flag to force
+	// getting all diffs, even if there is no change.
+	return m.diffList(k, schema, diff, d, true)
 }
 
 func (m schemaMap) diffString(
 	k string,
 	schema *Schema,
 	diff *terraform.InstanceDiff,
-	d *ResourceData) error {
+	d *ResourceData,
+	all bool) error {
 	var originalN interface{}
 	var os, ns string
 	o, n, _, _ := d.diffChange(k)
@@ -646,7 +668,7 @@ func (m schemaMap) diffString(
 		return fmt.Errorf("%s: %s", k, err)
 	}
 
-	if os == ns {
+	if os == ns && !all {
 		// They're the same value. If there old value is not blank or we
 		// have an ID, then return right away since we're already setup.
 		if os != "" || d.Id() != "" {