+++ CHANGELOG.md
@@ -42,6 +42,8 @@ BUG FIXES:
   * providers/aws: Disassociate EIP before destroying.
   * providers/aws: ELB treats subnets as a set.
   * providers/aws: Fix case where in a destroy/create tags weren't reapplied. [GH-464]
+  * providers/aws: Fix incorrect/erroneous apply cases around security group
+      rules. [GH-457]
   * providers/consul: Fix regression where `key` param changed to `keys. [GH-475]
 
 ## 0.3.0 (October 14, 2014)